+++ airflow/providers/google/cloud/utils/bigquery_get_data.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 from collections.abc import Iterator
 from logging import Logger
 
-from google.cloud.bigquery.table import Row
+from google.cloud.bigquery.table import Row, RowIterator
 
 from airflow.providers.google.cloud.hooks.bigquery import BigQueryHook
 
@@ -37,13 +37,16 @@ def bigquery_get_data(
 
     i = 0
     while True:
-        rows: list[Row] = big_query_hook.list_rows(
+        rows: list[Row] | RowIterator = big_query_hook.list_rows(
             dataset_id=dataset_id,
             table_id=table_id,
             max_results=batch_size,
             selected_fields=selected_fields,
             start_index=i * batch_size,
         )
+
+        if isinstance(rows, RowIterator):
+            raise TypeError("BigQueryHook.list_rows() returns iterator when return_iterator=False (default)")
 
         if len(rows) == 0:
             logger.info("Job Finished")